Add a virtual mode.
This mode works very similarly to the existing one, but executes on the
host instead of in an SGX enclave. This can be useful for development
and debugging, but provides no verifiable security guarantees about its
output.

When enabled, attested-fetch will link against the system's copy of
OpenSSL and Curl.